Split off path addition paint calls
